Feat: 알림 정보와 알림 로그의 독립성을 위한 데이터베이스 관계 변경으로 인한 수정#13

+++ src/main/java/com/notificationserver/application/port/in/dto/NotificationLogCountInDto.java
@@ -6,9 +6,9 @@ import lombok.Getter;
 @Getter
 @Builder
 public class NotificationLogCountInDto {
-	private Integer notificationLogCount;
+	private long notificationLogCount;
 
-	public static NotificationLogCountInDto getCount(Integer count) {
+	public static NotificationLogCountInDto getCount(long count) {
 		return NotificationLogCountInDto.builder()
 				.notificationLogCount(count)
 				.build();
